add POST for study leave
